% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microPopGut.R
\name{microPopGut}
\alias{microPopGut}
\title{microPopGut
runs the human colon version of microPop}
\usage{
microPopGut(
  numDays,
  time.step,
  transitTime,
  microbeNames,
  microbeNames.short,
  numStrains = 1,
  absorption = c(water = 3, SCFA = 9.6),
  pH.pars = list(pH.fixed = c(5.7, 6.2, 6.6), fix.pH = FALSE, pHLimit = TRUE),
  colon.pars = list(vol.litres = c(0.41, 0.98, 1.63), compartmentNames = c("prox",
    "trans", "dist")),
  bowel.movements = list(BM.duration.h = 15/60, frac.distal.emptied = 0.9, BMpd = 0,
    start.BM.time = list(7, c(7, 19), c(7, 15, 21))),
  init = list(C = 1, P = 1, B = 10, Acetate = 0.3606, Propionate = 0.1482, Butyrate =
    0.1762, W = 100),
  inflow = list(C = 50, P = 10, W = 1100, RS.frac = 0.78),
  meals = list(seed = 1, fluc.inflow = FALSE, fluc.subst.comp = FALSE, plotInflow =
    FALSE, saveInflowFig = FALSE, breakfast.start = 7, lunch.start = 13, dinner.start =
    19, meal.duration.h = 0.5, time.to.reach.colon.h = 7, gamma.mag = 1),
  mucin = list(mucin.rate = 5, RS.frac.mucin = 0.5, mucin.degrader = c("Bacteroides",
    "NoButyFibreDeg"), halfSat.mucin = 0.5, P.frac.mucin = 0.2, C.frac.mucin = 0.75),
  odeFunc = microPop::derivsDefault,
  oneStrainRandomParams = FALSE,
  microbeMolarMass = 113,
  bacCutOff = 1e-14,
  plotOptions = list(yLabel = "Concentration (g/L)", xLabel = "Time", plotFig = FALSE,
    sumOverStrains = FALSE, saveFig = FALSE, figType = "eps", figName = "microPopFig"),
  odeOptions = list(atol = 1e-06, rtol = 1e-06, method = "lsoda"),
  strainOptions = list(randomParams = c("halfSat", "yield", "maxGrowthRate",
    "pHtrait"), seed = 1, distribution = "uniform", percentTraitRange = 0, maxPHshift =
    0, applyTradeOffs = FALSE, tradeOffParams = NULL, paramsSpecified = FALSE,
    paramDataName = NULL),
  checkingOptions = list(checkMassConv = FALSE, balanceTol = 0.01, reBalanceStoichiom =
    FALSE, stoiTol = 0.1, checkForNegs = TRUE, negTol = -0.01, checkStoichiomBalance =
    TRUE),
  ...
)
}
\arguments{
\item{numDays}{number of days over which to run simulation}

\item{time.step}{simulation time step (in days)}

\item{transitTime}{time taken to travel through colon (d)}

\item{microbeNames}{named vector with names of microbial data frames}

\item{microbeNames.short}{named vector with shortened names for plotting}

\item{numStrains}{scalar if all groups have the same number of strains or a named vector specifying the number of strains in each microbial group.}

\item{absorption}{named vector with specific absorption rates (/d) for water and SCFA. Default is c('water'=3,'SCFA'=9.6)}

\item{pH.pars}{list controlling pH in the model. Default is list(pH.fixed=c(5.7,6.2,6.6),fix.pH=FALSE,pHLimit=TRUE),
\itemize{
\item    pH.fixed=c(5.7,6.2,6.6) pH values in prox, trans and distal compartments - used if fix.pH is TRUE
\item    fix.pH=FALSE Set to TRUE to keep pH fixed. If FALSE, pH varies with SCFA concentrations
\item    pHLimit=TRUE If TRUE then microbial growth is pH limited
}}

\item{colon.pars}{list of parameters describing colon. Default is list(vol.litres=c(0.41,0.98,1.63),compartmentNames=c('prox','trans','dist'))}

\item{bowel.movements}{List of parameters for modelling bowel movements. Default is list(BM.duration.h=15/60,frac.distal.emptied=0.90, BMpd=0,start.BM.time=list(7,c(7,19),c(7,15,21))
\itemize{
\item BM.duration.h time for bowel movement in hours
\item frac.distal.emptied fraction of the distal colon that is emptied by bowel movemen
\item BMpd number of bowel movements per day
\item start.BM.time list of times of bowel movements for 1, 2, or 3 BM per day
}}

\item{init}{list of initial values for carbohydrates (C), protein (P), microbes (B), Acetate, Propionate and Butyrate and water (W) in grams. Default values are: list(C=1,P=1,B=10,Acetate=0.3606,Propionate=0.1482,Butyrate=0.1762, W=100). SCFA initial values are based on concentrations of 60 mM, 20 mM, 20 mM for acetate, propionate and butyrate. The microbial mass is split evenly between the microbial groups. The initial conditions are same in each model compartment.}

\item{inflow}{list specifying the inflow of substrates (C, P) and water (W) in g/d, to the colon. Default values are: list(C=50,P=10,W=1100,RS.frac=0.78). The RS.frac is the fraction of RS in the carbohydrate i.e. RS/(RS+NSP) assuming RS+NSP=C}

\item{meals}{list of parameters describing meals. Default is list(seed=1,fluc.inflow=FALSE,fluc.subst.comp=FALSE,plotInflow=FALSE,saveInflowFig=FALSE,breakfast.start=7,lunch.start=13,dinner.start=19,meal.duration.h=0.5,time.to.reach.colon.h=7,gamma.mag=1)
\itemize{
\item seed seed for random number generator used for stochastic meal composition only (fluc.subst.comp=TRUE)
\item fluc.inflow FALSE if incoming substrate is constant i.e. TRUE means meals
\item fluc.subst.comp FALSE change to TRUE for substrate composition varying randomly
\item plotInflow FALSE change to TRUE to plot the incoming substrate
\item saveInflowFig FALSE Change to TRUE to save inflow figs
\item breakfast.start time breakfast is eaten (24 h clock)
\item lunch.start time lunch is eaten (24 h clock)
\item dinner.start time dinner is eaten (24 h clock)
\item meal.duration.h time taken to eat a meal (in hours)
\item time.to.reach.colon.h time taken to move through stomach and small intestine etc (h)
\item gamma.mag (positive number) this determines the level of variation in diet fluctuations (1.1 will give bigger fluctuations, 0.9 smaller).
}}

\item{mucin}{list of parameters describing mucins. Default is list(mucin.rate=5,RS.frac.mucin=0.5,mucin.degrader=c('Bacteroides','NoButyFibreDeg'),halfSat.mucin=0.5,P.frac.mucin=0.2,C.frac.mucin=0.75)
\itemize{
\item mucin.rate  rate of mucin (g/d) secreted into colon
\item RS.frac.mucin fraction of C in mucin that is RS
\item mucin.degrader microbes that break down mucin
\item halfSat.mucin half saturation constant (g/l of B) for mucin breakdown
\item P.frac.mucin fraction of mucin that is protein
\item C.frac.mucin fraction of mucin that is carbohydrate
}}

\item{odeFunc}{The function the ODE solver in microPopModel() will use - the default is derivsDefault provided by the package but if the user wants to make significant changes a new ODE function file can be used. See ?derivsDefault}

\item{oneStrainRandomParams}{Logical to allow randomization of parameters even if there is only one strain. The default is FALSE which means that if numStrains=1 then the group params are used; if numStrains>1 then the parameters are automatically randomised according to info given in strainOptions. If oneStrainRandomParams=TRUE then even if there is only one strain its parameters will be randomised according to info given in strainOptions.}

\item{microbeMolarMass}{Scalar. Mass of 1 mole of microbes - default is 113g/mol (Batstone et al., 2002)}

\item{bacCutOff}{Scalar. Amount of bacteria below which the bacteria are considered to have left the system and can't grow, default =1e-14. If this is set to zero then bacteria will always be able to grow again as zero is never reached.}

\item{plotOptions}{List containing instructions for plotting: Default is list(plotFig=TRUE, sumOverStrains=FALSE, saveFig=FALSE, figType='eps', figName='microPopFig', yLabel='Concentration (g/L)', xLabel='Time').\cr
To turn off plot generation set plotFig=FALSE. If there are multiple strains these are all plotted if sumOverStrains=FALSE, otherwise they will be summed over each group. To save plot, saveFig=TRUE, figType (format) can be 'eps','png', 'pdf' or 'tiff' and is specified in figType (string), the name is figName (string) to which the string 'Microbes' or 'Resources' will be added for the respective plots.}

\item{odeOptions}{List containing instructions for the ODE solver ('deSolve'). Default: list('atol'=1e-6,'rtol'=1e-6,'method'='lsoda'). See ?ode for more details.}

\item{strainOptions}{List containing instructions for specifying strain parameters. Default: list(randomParams=c('halfSat', 'yield', 'maxGrowthRate', 'pHtrait'), seed=1, distribution='uniform', percentTraitRange=0, maxPHshift=0., applyTradeOffs=FALSE, tradeOffParams=NULL, paramsSpecified=FALSE, paramDataName=NULL).
\itemize{
\item randomParams (vector) specifying which parameters need to be stochastically generated.
\item seed (number) seed for random number generator.
\item distribution (string) - either 'uniform' or 'normal' specifying the shape of the distribution from which to draw the random strain parameters.
\item percentTraitRange (number) this is the percentage either side of the group parameter value which the strain parameter may range e.g. if percentTraitRange=10 then range is 0.9x to 1.1x for group mean x.
\item maxPHshift (number) pH units to range over.
\item applyTradeOffs (logical) to trade off \verb{good' and }bad' parameter values.
\item tradeOffParams (vector of two strings) - parameters to trade off against each other. Note that pHtrait can not be traded off as whether this trait is good or bad depends on the environmental pH.
\item paramsSpecified (logical) TRUE if strain parameters are read in from a file (whose name is specified in paramDataName). The file must have colnames c(strainName, paramName, paramVal, paramUnit, resource,path) and where strainName is in format 'groupName.i' where i is the strain number.
}}

\item{checkingOptions}{(List) Default is list(checkMassConv=FALSE, balanceTol=1e-2, reBalanceStoichiom=FALSE, stoiTol=0.1, checkForNegs=TRUE, negTol=-1e-2).
\itemize{
\item checkMassConv=TRUE checks for mass conservation in the ODE solver with a tolerance of 'balanceTol' (default is FALSE).
\item reBalanceStoichiom will check the mass balance of the stoichiometries on every metabolic path and rebalance if these are not conserving mass within a tolerance of stoiTol (a warning message will be issued). Rebalancing will only affect the final solution if the pathway contains only essential resources (Rtype 'Se') and microbial biomass is a product (Rtype 'Pb').
\item checkForNegs If TRUE the function checkSolution is called and the solution for each variable, x, is checked for negative values that are greater in magnitude than negTol*max(x). If negative values occur then the solution is incorect and either the problem is incorrectly specified or the tolerances in the ODE solver need to be smaller.
}}

\item{...}{Add your own inputs}
}
\value{
The output is a list containing lists called 'solution', 'params' and myPars. These lists have 3 components (one for each model compartment (i.e. the proximal, transverse and distal colon). Solution[[n]] (for n=1,2,or 3 depending on compartment) is a matrix where rows are points in time and the columns are the state variables. params[[n]] contains all the information needed to run the model. Use names(out$params[[1]]) to see contents.
}
\description{
microPopGut
runs the human colon version of microPop
}
