% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRateFunctions.R
\docType{data}
\name{newRateFuncs}
\alias{newRateFuncs}
\title{newRateFuncs for microPopGut}
\format{
An object of class \code{list} of length 11.
}
\usage{
newRateFuncs
}
\description{
List of functions that are used by the ODE solver
these functions can be changed by the user but all must be listed.
}
\details{
rateFuncsDefault=list(pHFunc=pHFuncDefault, pHLimFunc=pHLimFuncDefault, extraGrowthLimFunc=extraGrowthLimFuncDefault, growthLimFunc=growthLimFuncDefault, combineGrowthLimFunc=combineGrowthLimFuncDefault, uptakeFunc=uptakeFuncDefault, productionFunc=productionFuncDefault, combinePathsFunc=combinePathsFuncDefault, massBalanceFunc=massBalanceFuncDefault, entryRateFunc=entryRateFuncDefault, removalRateFunc=removalRateFuncDefault)

note that in these functions, the parms list is intrinsic to microPop whereas the myPars list is defined by the inputs to microPopGut() and then added to parms
}
\keyword{datasets}
