% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHFuncMPG.R
\name{pHFuncMPG}
\alias{pHFuncMPG}
\title{pH Function for microPopGut}
\usage{
pHFuncMPG(time, parms, stateVarValues)
}
\arguments{
\item{time}{(scalar). The current time point in the ODE solver.}

\item{parms}{List which contains all information required by the ODE solver}

\item{stateVarValues}{State vector (resources and microbes) (with names)}
}
\value{
(scalar) pH at the given time
}
\description{
Return the value of pH in pH units
}
