% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPG.R
\name{plotMPG}
\alias{plotMPG}
\title{plotMPG}
\usage{
plotMPG(m.out)
}
\arguments{
\item{m.out}{is the output of microPopGut(). It is a list containing the following objects: solution, params, myPars, etc}
}
\description{
Basic plot of the model results
}
